pyproject.toml
[project]
name = "zettelkasten-auto"
version = "0.1.0"
description = "Convert source notes into atomic Zettelkasten notes using LLM extraction and semantic linking"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "chromadb>=0.4.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=2.2.0",
    "google-generativeai>=0.5.0",
    "requests>=2.28.0",
]

[project.scripts]
zettel = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
